requirements.txt
aiomysql==0.2.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Django==5.1.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
Flask==3.1.0
Flask-SocketIO==5.4.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
orjson==3.10.11
passlib==1.7.4
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.17
python-socketio==5.11.4
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.41.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.0
Werkzeug==3.1.3
wsproto==1.2.0